+++ src/api/restful/sell/listing/index.ts
@@ -11,17 +11,10 @@ export default class Listing extends Restful {
   /**
    * This call gives Partners the ability to create an eBay draft of a item for their seller using information from their site.
    *
-   * @param contentLanguage Use this header to specify the natural language of the seller.
-   * @param marketplaceId se this header to specify an eBay marketplace ID.
    * @param data The ItemDraft
    */
-  public createItemDraft(contentLanguage: string, marketplaceId: string, data?: any) {
-    return this.post(`/item_draft/`, data, {
-      headers: {
-        'Content-Language': contentLanguage,
-        'X-EBAY-C-MARKETPLACE-ID': marketplaceId,
-      }
-    })
+  public createItemDraft(data?: any) {
+    return this.post(`/item_draft/`, data)
   }
 
 }