+++ src/api/restful/sell/recommendation/index.ts
@@ -15,8 +15,6 @@ export default class Recommendation extends Restful {
    * The find method returns recommendations and information that sellers can use to optimize their listing
    * configurations.
    *
-   * @param marketplaceId Use this header to specify the eBay marketplace where you list the items for which you want
-   *     to get recommendations.
    * @param filter Provide a list of key-value pairs to specify the criteria you want to use to filter the response.
    * @param limit Use this query parameter to set the maximum number of ads to return on a page from the paginated
    *     response. Default: 10 Maximum: 500
@@ -24,8 +22,7 @@ export default class Recommendation extends Restful {
    *     paginated response.
    * @param body FindListingRecommendationRequest
    */
-  public findListingRecommendations(
-    marketplaceId: string, {filter, limit, offset}: { filter?: string, limit?: number, offset?: number } = {},
+  public findListingRecommendations({filter, limit, offset}: { filter?: string, limit?: number, offset?: number } = {},
     body?: FindListingRecommendationRequest
   ) {
     return this.post(`/find`, {
@@ -34,9 +31,6 @@ export default class Recommendation extends Restful {
         filter,
         limit,
         offset
-      },
-      headers: {
-        'X-EBAY-C-MARKETPLACE-ID': marketplaceId
       }
     });
   }