+++ src/api/traditional/index.ts
@@ -1,6 +1,6 @@
 import {stringify} from 'qs';
 import Api from '../';
-import {EBayIAFTokenExpired, handleEBayError} from '../../errors';
+import {EBayIAFTokenExpired, EBayIAFTokenInvalid, handleEBayError} from '../../errors';
 import {ClientAlerts, Finding, Shopping, Trading, TraditionalApi} from '../../types';
 import ClientAlertsCalls from './clientAlerts';
 import {Fields} from './fields';
@@ -119,15 +119,8 @@ export default class Traditional extends Api {
     return service;
   }
 
-  // TODO
   public createBusinessPolicyManagementApi() {
-    const api = {
-      headers: (_: string, accessToken?: string) => ({
-        ...(accessToken && {'X-EBAY-SOA-SECURITY-IAFTOKEN': accessToken})
-      })
-    };
-
-    return api;
+    throw new Error('Important! This API is deprecated and will be decommissioned on January 31, 2022. We recommend that you migrate to the fulfillment_policy, payment_policy, and return_policy resources of the Account API to set up and manage all of your fulfillment, payment, and return business policies.')
   }
 
   private createXMLRequest = (callName: string, api: TraditionalApi) => async (fields: Fields, opts: Options) => {
@@ -135,9 +128,9 @@ export default class Traditional extends Api {
 
     try {
       return await this.request(options, api, callName, fields);
-    } catch (error) {
+    } catch (error: any) {
       // Try to refresh the token.
-      if (error.name === EBayIAFTokenExpired.name && this.config.autoRefreshToken) {
+      if (this.config.autoRefreshToken && (error.name === EBayIAFTokenExpired.name || error.name === EBayIAFTokenInvalid.name)) {
         return await this.request(options, api, callName, fields, true);
       }
 
@@ -148,7 +141,7 @@ export default class Traditional extends Api {
   private async request(options: Options, api: TraditionalApi, callName: string, fields: Fields, refreshToken = false) {
     try {
       if (refreshToken) {
-        await this.auth.OAuth2.refreshAuthToken();
+        await this.auth.OAuth2.refreshToken();
       }
 
       const config = this.getConfig(api, callName, options);
@@ -162,29 +155,29 @@ export default class Traditional extends Api {
 
   private getConfig(api: TraditionalApi, callName: string, options: Options) {
     const eBayAuthToken = this.auth.authNAuth.eBayAuthToken;
-    const accessToken = this.auth.OAuth2.accessToken;
-    const useIafToken = (!eBayAuthToken || accessToken && options.useIaf);
+    const userAccessToken = this.auth.OAuth2.getUserAccessToken();
+    const useIaf = (!eBayAuthToken || userAccessToken && options.useIaf);
 
     return {
       ...options,
       xmlns: api.xmlns,
       endpoint: api.endpoint[this.config.sandbox ? 'sandbox' : 'production'],
       headers: {
-        ...api.headers(callName, accessToken && useIafToken ? accessToken : undefined),
+        ...api.headers(callName, userAccessToken && useIaf ? userAccessToken : undefined),
         ...options.headers
       },
-      ...(eBayAuthToken && !useIafToken && {
+      ...(eBayAuthToken && !useIaf && {
         eBayAuthToken
       })
     };
   }
 
-  private createTraditionalXMLApi<T>(api: TraditionalApi): T {
-    const service: any = {};
-    Object.keys(api.calls).forEach((callName: string) => {
-      service[callName] = this.createXMLRequest(callName, api);
+  private createTraditionalXMLApi<T>(traditionalApi: TraditionalApi): T {
+    const api: any = {};
+    Object.keys(traditionalApi.calls).forEach((callName: string) => {
+      api[callName] = this.createXMLRequest(callName, traditionalApi);
     });
 
-    return service as T;
+    return api as T;
   }
 }