+++ test/api/oAuth2.spec.ts
@@ -49,7 +49,7 @@ describe('OAuth2', () => {
       const oAuth2 = new OAuth2(config, req);
       try {
         oAuth2.generateAuthUrl();
-      } catch (e) {
+      } catch (e: any) {
         expect(e.message).to.equal('RuName is required.')
       }
     });
@@ -68,7 +68,7 @@ describe('OAuth2', () => {
       const oAuth2 = new OAuth2(config, req);
       oAuth2.setClientToken(cred)
 
-      const token = await oAuth2.getClientAccessToken();
+      const token = await oAuth2.getApplicationAccessToken();
 
       expect(token).to.equal('access_token');
     });
@@ -80,8 +80,8 @@ describe('OAuth2', () => {
       });
 
       try {
-        await oAuth2.getClientAccessToken();
-      } catch (e) {
+        await oAuth2.getApplicationAccessToken();
+      } catch (e: any) {
         expect(e.name).to.equal('error');
       }
     });
@@ -89,7 +89,7 @@ describe('OAuth2', () => {
     it('refresh the client access token', async () => {
       const oAuth2 = new OAuth2(config, req);
 
-      const token = await oAuth2.getClientAccessToken();
+      const token = await oAuth2.getApplicationAccessToken();
       expect(token).to.equal('new_access_token');
     });
 
@@ -112,7 +112,7 @@ describe('OAuth2', () => {
 
       try {
         await oAuth2.getToken('code');
-      } catch (e) {
+      } catch (e: any) {
         expect(e.name).to.equal('error');
       }
     });
@@ -130,8 +130,8 @@ describe('OAuth2', () => {
     it('Throws error if appId is not defined', async () => {
       try {
         const oAuth2 = new OAuth2({...config, appId: ''}, req);
-        await oAuth2.refreshClientToken()
-      } catch (e) {
+        await oAuth2.obtainApplicationAccessToken()
+      } catch (e: any) {
         expect(e.message).to.equal('Missing App ID (Client Id)')
       }
     });
@@ -139,8 +139,8 @@ describe('OAuth2', () => {
     it('Throws error if appId is not defined', async () => {
       try {
         const oAuth2 = new OAuth2({...config, certId: ''}, req);
-        await oAuth2.refreshClientToken()
-      } catch (e) {
+        await oAuth2.obtainApplicationAccessToken()
+      } catch (e: any) {
         expect(e.message).to.equal('Missing Cert Id (Client Secret)')
       }
     });
@@ -149,9 +149,9 @@ describe('OAuth2', () => {
       const oAuth2 = new OAuth2(config, req);
 
       try {
-        await oAuth2.refreshAuthToken()
-      } catch (e) {
-        expect(e.message).to.equal('Failed to refresh the token. Token is not set.')
+        await oAuth2.refreshUserAccessToken()
+      } catch (e: any) {
+        expect(e.message).to.equal('Failed to refresh the user access token. Token or refresh_token is not set.')
       }
     })
 
@@ -159,8 +159,8 @@ describe('OAuth2', () => {
       const oAuth2 = new OAuth2(config, {...req, postForm: sinon.stub().throws('error')});
       oAuth2.setCredentials(cred);
       try {
-        await oAuth2.refreshAuthToken();
-      } catch (e) {
+        await oAuth2.refreshUserAccessToken();
+      } catch (e: any) {
         expect(e.name).to.equal('error');
       }
     });
@@ -169,8 +169,8 @@ describe('OAuth2', () => {
       const oAuth2 = new OAuth2(config, req);
       try {
         await oAuth2.refreshToken();
-      } catch (e) {
-        expect(e.message).to.equal('To refresh a Token a client token or user access token must be already set.');
+      } catch (e: any) {
+        expect(e.message).to.equal('Missing credentials. To refresh a token an application access token or user access token must be already set.');
       }
     });
 
@@ -207,8 +207,8 @@ describe('OAuth2', () => {
       const refreshAuthToken = sinon.stub();
       oAuth2.on('refreshAuthToken', refreshAuthToken);
 
-      return oAuth2.refreshAuthToken().then(() => {
-        expect(oAuth2.accessToken).equal('new_access_token');
+      return oAuth2.refreshUserAccessToken().then(() => {
+        expect(oAuth2.getUserAccessToken()).equal('new_access_token');
         // tslint:disable-next-line:no-unused-expression
         expect(refreshAuthToken.called).to.be.true;
         expect(refreshAuthToken.args[0][0].access_token).to.equal('new_access_token');